refactor: move save collection student method
It's now in the SdcSchoolCollectionStudentService.

squashable
